requirements.txt
torch==2.2.2
torchvision==0.17.2
numpy
pillow
spacy
tqdm
matplotlib
nltk
pandas
streamlit